backend/pyproject.toml
[project]
name = "holmes-backend"
version = "0.1.0"
description = "Holmes legal intelligence platform backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "sse-starlette>=2.2.0",
    "google-cloud-storage>=2.19.0",
    "google-adk>=1.22.0",
    "greenlet>=3.3.0",
    "PyJWT>=2.10.0",
    "cryptography>=43.0.0",
    "python-multipart>=0.0.22",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.13",
    "mypy>=1.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # allow Depends() in default arguments (FastAPI pattern)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
